split the models in core over multiple modules
