+++ app/backend/src/couchers/models/__init__.py
@@ -1,1 +1,2 @@
 from couchers.models.core import *
+from couchers.models.messaging import *
